:bug: Fix: 서버의 반환값에 맞게 데이터 타입 수정

+++ app/src/main/java/com/example/snack4diet/api/editFood/EditFood.kt
@@ -2,7 +2,7 @@ package com.example.snack4diet.api.editFood
 
 data class EditFood(
     val baseNutrition: BaseNutrition,
-    val foodId: Int,
+    val foodId: Long,
     val name: String,
     val userId: Long
 )
